Don't overwrite rectified_roi_dirty flag
This fixes a problem with the following sequence:

1. fromCameraInfo is called with ROI A.  | rectified_roi_dirty = true
2. rectifiedRoi is called                | rectified_roi_dirty = false
3. fromCameraInfo is called with ROI B.  | rectified_roi_dirty = true
4. fromCameraInfo is called with ROI B.  | rectified_roi_dirty = false

The bug is that rectified_roi_dirty was incorrectly set to `false` by the last
line. If now rectifiedRoi is called again, the cached rectified_roi for
ROI A will be returned, but it should be recalculated based on ROI B.

This commit fixes that.